Refactor duplicated utility functions and convert frontend to ESM.
- Extracted shared utility functions (getRunStats, getColor, escapeHtml, capitalize, formatTestName) into eval-view/utils.js.
- Converted eval-view/utils.js to use ESM exports.
- Updated eval-view/dashboard.js and eval-view/landing.js to use ESM imports.
- Updated eval-view/index.html and eval-view/dashboard.html to use type="module" for scripts.
- Consolidated and improved utility implementations.
