+++ eval-view/dashboard.html
@@ -41,7 +41,7 @@
     </div>
   </div>
 
-  <script src="dashboard.js"></script>
+  <script type="module" src="dashboard.js"></script>
 </body>
 
 </html>
