+++ eval-view/index.html
@@ -108,6 +108,6 @@
     </div>
   </div>
 
-  <script src="landing.js"></script>
+  <script type="module" src="landing.js"></script>
 </body>
 </html>
