+++ eval-view/landing.js
@@ -1,3 +1,5 @@
+import { getRunStats, getColor, escapeHtml, capitalize } from './utils.js';
+
 let allTestData = {}; // Cache all test data by testID
 let currentTab = 'overview';
 let currentScenarioFilter = 'all';
@@ -619,14 +621,6 @@ function calculateGroupTotalStats(results, groupType) {
     return { passed, total };
 }
 
-function getRunStats(checks) {
-    if (!checks || !checks.length) return { rate: 0, passed: 0, total: 0 };
-    const passed = checks.filter(c => c.passed).length;
-    const total = checks.length;
-    const rate = Math.round((passed / total) * 100);
-    return { rate, passed, total };
-}
-
 function getSortedTestIds() {
     // Return only SELECTED tests, sorted by date
     return Array.from(selectedTestIds).sort((a, b) => {
@@ -634,24 +628,4 @@ function getSortedTestIds() {
         if (!allTestData[a] || !allTestData[b]) return 0;
         return new Date(allTestData[b].timestamp) - new Date(allTestData[a].timestamp);
     });
-}
-
-function capitalize(s) {
-    return s.charAt(0).toUpperCase() + s.slice(1);
-}
-
-function getColor(percentage) {
-    if (percentage >= 90) return 'var(--accent-success)';
-    if (percentage >= 50) return '#dbab09';
-    return 'var(--accent-failure)';
-}
-
-function escapeHtml(text) {
-    if (!text) return text;
-    return text
-        .replace(/&/g, "&amp;")
-        .replace(/</g, "&lt;")
-        .replace(/>/g, "&gt;")
-        .replace(/"/g, "&quot;")
-        .replace(/'/g, "&#039;");
 }
